practical8/requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async support
asyncio-subprocess>=0.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0

# Logging
structlog>=23.0.0